Add a debug GUI for editing the player party
This adds a little GUI for editing the current party, similar to what you can do with Pkhex.

The idea is to be able to quickly change certain values of party Pokémon to make testing a bit easier.

It is _not_ meant to replace Pkhex and I don't want to treat this as a regular feature of the bot and support it, so it is purposefully hidden in a Debug menu.